fix: fixed trimming name and separated color picker into util file
