+++ frontend/appflowy_flutter/lib/util/color_picker/color_picker_service.dart
@@ -1,0 +1,7 @@
+import 'dart:ui';
+
+/// Abstract random color picker as a service to implement dependency injection.
+abstract class ColorPickerService {
+  Color generateRandomNameColor(String name) =>
+      throw UnimplementedError('generateNameColor() has not been implemented.');
+}
